Comparing graph generation techniques speed.
Added a function graphgen2 that uses add_vertex in igraph one by one.
This is pretty slow. Comparing it with graphgen which creates a huge
list of vertices.

It might be more efficient to yield an iterator that
doesn't store N**2 vertices in a list. graphgen takes about 176 seconds.
I remember graphgen2 taking about 125 seconds for N = 300. I'm testing
it right now.