Added correlation plotting functions, absnormal, etc
I added two functions:

1.  plot_correlations. This plots the correlations and covariance of
bus1,bus1 and correctly sets the the plot title and legend.
1. absnormal. This simply calls the absolute value of the normal
distribution from the numpy.random package. Another test for the
negative correlations of the Busemann function.
1.  gamma_fit, gamma_pdf were functions added in an attempt to "fit" the
pdf of a gamma to the busemann density of a uniform. Clearly there is
some sort of discontinuity at $1$.

I updated import_from_file to save the graph to the file as well.
Shelve appears to pickly it correctly.