Tried to implement geodesics
I tried and failed since `get_all_shortest_paths` is not going to work
for us. We have to instead get all the times possible and work backwards
to discover the shortest path in the usual dynamic programming way.
Perhaps this will reduce the amount of effort anyway since originally I
was calling

    g.shortest_paths_dijkstra
    g.get_shortest_paths

twice.