New functionality to compare covariance plot
Added an option called x_range to plot_correlation
that is helpful to compare covariances of two different weight
distributions. This allows you to manually specify the x range
over which the covariance or correlation will be plotted
so that you can keep it consistent over two calls to plot_correlation
with pairs of busemann functions from different distributions.

Also modified plot_correlation to return the lines it plots so that you
can easily modify it from outside.