New functionality to compare covariance plot
Added an option called x_range to plot_correlation
that is helpful to compare covariances of two different weight
distributions. This allows you to manually specify the x range
over which the covariance or correlation will be plotted
so that you can keep it consistent over two calls to plot_correlation
with pairs of busemann functions from different distributions.

Also modified plot_correlation to return the lines it plots so that you
can easily modify it from outside.

Fixed the savefig title so that correlation or covariance comes first.
Originally it said that 'busemann' first, and this makes it hard to keep
track of files in the directory. Now images are grouped together by name
(correlation exponential N=200 samples=500)

To plot_shape_pyplot I also added

    exp_plot_options
    plot_options

since the exponential universal curve is plotted as a contour plot, and
the limit shape interface is plotted as a scatterplot. They take
different options for linewidth and so on, so I did this.

I also made some changes to plot_geodesics so that the plots are easier
to see. In addition to different colors, now geodesics can be
visualized with different shapes: square,circle,triangle. Somehow,
even though the igraph library says that triangles are supported, they
did not seem to plot. Now, plot_geodesics returns the
plots, the joint_color_vector and the joint_shape_vector