Working on exponential martingale for the polymer
Based on Song-Zhou second moment computations. Seeing if the
exponential martingale for the polymer goes to zero.

Seems like the 2nd moment blows up, and the exp martingale converges
to 0.

We checked polymer.pyx and it seems to work with vertex weights. We had
to remove a 1/2 from the Z definition.

edge weights can be added at some other point. We added exp_martingale.py.